Refactor test/e2e to separate kubernetes from efs tests
