+++ test/e2e/efs/README.md
@@ -1,0 +1,1 @@
+Tests EFS-specific functionality like encryption in transit.
