+++ test/e2e/kubernetes/README.md
@@ -1,0 +1,3 @@
+Tests standard functionality
+https://kubernetes-csi.github.io/docs/functional-testing.html#functional-testing
+(using the "Importing the E2E test suite as a library" approach)
