+++ test/e2e/kubernetes/e2e.go
@@ -1,0 +1,104 @@
+package e2e
+
+import (
+	"fmt"
+
+	"github.com/onsi/ginkgo"
+	v1 "k8s.io/api/core/v1"
+	storagev1 "k8s.io/api/storage/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/sets"
+	"k8s.io/kubernetes/test/e2e/framework"
+	storageframework "k8s.io/kubernetes/test/e2e/storage/framework"
+	"k8s.io/kubernetes/test/e2e/storage/testsuites"
+	"k8s.io/kubernetes/test/e2e/storage/utils"
+)
+
+var (
+	FileSystemId string
+)
+
+type efsDriver struct {
+	driverInfo storageframework.DriverInfo
+}
+
+var _ storageframework.TestDriver = &efsDriver{}
+
+// TODO implement Inline (unless it's redundant)
+// var _ testsuites.InlineVolumeTestDriver = &efsDriver{}
+var _ storageframework.PreprovisionedPVTestDriver = &efsDriver{}
+var _ storageframework.DynamicPVTestDriver = &efsDriver{}
+
+func InitEFSCSIDriver() storageframework.TestDriver {
+	return &efsDriver{
+		driverInfo: storageframework.DriverInfo{
+			Name:            "efs.csi.aws.com",
+			SupportedFsType: sets.NewString(""),
+			Capabilities: map[storageframework.Capability]bool{
+				storageframework.CapPersistence: true,
+				storageframework.CapExec:        true,
+				storageframework.CapMultiPODs:   true,
+				storageframework.CapRWX:         true,
+			},
+		},
+	}
+}
+
+func (e *efsDriver) GetDriverInfo() *storageframework.DriverInfo {
+	return &e.driverInfo
+}
+
+func (e *efsDriver) SkipUnsupportedTest(storageframework.TestPattern) {}
+
+func (e *efsDriver) PrepareTest(f *framework.Framework) (*storageframework.PerTestConfig, func()) {
+	cancelPodLogs := utils.StartPodLogs(f, f.Namespace)
+
+	return &storageframework.PerTestConfig{
+			Driver:    e,
+			Prefix:    "efs",
+			Framework: f,
+		}, func() {
+			cancelPodLogs()
+		}
+}
+
+func (e *efsDriver) CreateVolume(config *storageframework.PerTestConfig, volType storageframework.TestVolType) storageframework.TestVolume {
+	return nil
+}
+
+func (e *efsDriver) GetPersistentVolumeSource(readOnly bool, fsType string, volume storageframework.TestVolume) (*v1.PersistentVolumeSource, *v1.VolumeNodeAffinity) {
+	pvSource := v1.PersistentVolumeSource{
+		CSI: &v1.CSIPersistentVolumeSource{
+			Driver:       e.driverInfo.Name,
+			VolumeHandle: FileSystemId,
+		},
+	}
+	return &pvSource, nil
+}
+
+func (e *efsDriver) GetDynamicProvisionStorageClass(config *storageframework.PerTestConfig, fsType string) *storagev1.StorageClass {
+	parameters := map[string]string{
+		"provisioningMode": "efs-ap",
+		"fileSystemId":     FileSystemId,
+		"directoryPerms":   "777",
+	}
+
+	generateName := fmt.Sprintf("efs-csi-dynamic-sc-test1234-")
+
+	defaultBindingMode := storagev1.VolumeBindingImmediate
+	return &storagev1.StorageClass{
+		ObjectMeta: metav1.ObjectMeta{
+			GenerateName: generateName,
+		},
+		Provisioner:       "efs.csi.aws.com",
+		Parameters:        parameters,
+		VolumeBindingMode: &defaultBindingMode,
+	}
+}
+
+var _ = ginkgo.Describe("[efs-csi] EFS CSI kubernetes", func() {
+	driver := InitEFSCSIDriver()
+	ginkgo.Context(storageframework.GetDriverNameWithFeatureTags(driver), func() {
+		storageframework.DefineTestSuites(driver, testsuites.CSISuites)
+	})
+})
